Admin Profile - Working With Update Feature (Part - 4)

+++ app/Traits/FileUploadTrait.php
@@ -7,7 +7,7 @@ use File;
 
 trait FileUploadTrait
 {
-    function uploadImage(Request $request, string $inputName, string $oldPath = null, string $path = '/uploads') : ?string
+    function uploadImage(Request $request, string $inputName, ?string $oldPath = null, string $path = '/uploads') : ?string
     {
         if($request->hasFile($inputName)){
             $image = $request->{$inputName};
@@ -17,7 +17,9 @@ trait FileUploadTrait
             $image->move(public_path($path), $imageName);
 
             // Delete previous image from storage
-            if($oldPath && File::exists(public_path($oldPath))){
+            $exculudedFolder = '/default';
+
+            if($oldPath && File::exists(public_path($oldPath)) && strpos($oldPath, $exculudedFolder) !== 0){
                 File::delete(public_path($oldPath));
             }
